tidy up: decomposed in sub packages
