+++ demos/demo_zmx.py
@@ -34,7 +34,7 @@ from pyrateoptics.core.localcoordinates import LocalCoordinates
 from pyrateoptics.core.material_isotropic import ConstantIndexGlass
 from pyrateoptics.core.globalconstants import standard_wavelength, numerical_tolerance
 from pyrateoptics.core.ray import RayBundle
-from pyrateoptics.core.zmx import ZMXParser
+from pyrateoptics.io.zmx import ZMXParser
 
 from pyrateoptics.core import raster
