+++ pyrateoptics/analysis/optical_element_analysis.py
@@ -25,7 +25,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
 import numpy as np
-from log import BaseLogger
+from ..core.log import BaseLogger
 
 class OpticalElementAnalysis(BaseLogger):
