+++ pyrateoptics/analysis/optical_system_analysis.py
@@ -25,7 +25,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 """
 
 import numpy as np
-from log import BaseLogger
+from ..core.log import BaseLogger
 from ray_analysis import RayBundleAnalysis
 import matplotlib.pyplot as plt
