+++ pyrateoptics/analysis/surfShape_analysis.py
@@ -26,7 +26,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import numpy as np
 import matplotlib.pyplot as plt
-from log import BaseLogger
+from ..core.log import BaseLogger
 
 class ShapeAnalysis(BaseLogger):
