+++ pyrateoptics/core/material.py
@@ -26,7 +26,7 @@ Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 
 import numpy as np
 import math
-from optimize import ClassWithOptimizableVariables
+from ..optimize.optimize import ClassWithOptimizableVariables
 import scipy.linalg as sla
 
 from globalconstants import standard_wavelength
