--- pyrateoptics/core/stub.py
@@ -1,29 +1,0 @@
-"""
-Pyrate - Optical raytracing based on Python
-
-Copyright (C) 2017 Moritz Esslinger <moritz.esslinger@web.de>
-               and Johannes Hartung <j.hartung@gmx.net>
-               and     Uwe Lippmann <uwe.lippmann@web.de>
-               and    Thomas Heinze <t.heinze@fn.de>
-
-This program is free software; you can redistribute it and/or
-modify it under the terms of the GNU General Public License
-as published by the Free Software Foundation; either version 2
-of the License, or (at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program; if not, write to the Free Software
-Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston,
-MA 02110-1301, USA.
-"""
-
-__author__ = 'Thomas Heinze'
-
-def func(val):
-    """Test stub."""
-    return val + 1
