+++ pyrateoptics/sampling2d/raster.py
